+++ packages/style-value-types/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Style Value Types adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.0.0] 2018-06-22
+
+### Removed
+
+- Removed new `combo` in favour of a single unified `complex` value type that can handle mixed number and color strings.
+
 ## [2.0.1] 2018-06-21
 
 ### Fixed